chore: Initial Sketch of how a hyperdrive might be implemented
Doesn't currently work. First issue with the implementation is that
Hyperdrive retrieves the header from the underlying hyperbee on
readying. This is an issue in two potential ways, first because if a
header is automatically added in a normal workflow, autobase wont allow
that as it would happen outside of the `apply` function. Second, when
the hyperdrive requests the header it specifies that the request for the
block shouldn't wait. Currently the autobase `AutocoreSession` throws an
error when there isn't a block and doesn't support a wait version.

Another issue currently is that even faking the functionality to not
wait for a block with a proxy still throws a different error when
hyperdrive calls `this.db.core.core.compact`. `.core` doesn't exist on
`this.db.core` because it's a `AutocoreSession`.