fix: Defer creating Hyperdrive until `apply()` call
Previously Hyperdrive was throwing an error because the header block
wasn't created yet and AutocoreSession didn't support the `{ wait:
false }` option. Instead of using a proxy to emulate a hyperbee, this
approach defers creating the hyperdrive until in the `apply()` function
so all initializing of the hyperbee is not an issue. This cached version
is used thereafter.

`{ prefetch: false }` was required in `.get(path, opts)` because the
prefetch a) uses the hypercore `core.core` which isn't available on
`AutocoreSession` b) uses `core.download()` which also isn't available.
If these are implemented in the future, prefetching would be nice.